change-request-system-kev-devmethod-frontend-crp-ui/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
bleach==6.2.0
bokeh==3.6.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
fonttools==4.56.0
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
linkify-it-py==2.0.3
llamaapi==0.1.36
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
panel==1.6.1
param==2.2.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
propcache==0.3.0
pycparser==2.22
pyparsing==3.2.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyviz_comms==3.0.4
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
urllib3==2.3.0
webencodings==0.5.1
xyzservices==2025.1.0
yarl==1.18.3
